Attempt at generalising elevated support
Refactor puppet-server provisioner to use it.
